Separate out CVE downloading into separate app, TODO passing arguments
