+++ vulnfeeds/.dockerignore
@@ -1,5 +1,5 @@
 osv_interm_output
 osv_output
-output2
+cve_jsons
 parts
 .idea
