+++ vulnfeeds/.gitignore
@@ -1,6 +1,6 @@
 osv_interm_output
 pypi_links.json
 pypi_versions.json
-output2
+cve_jsons
 osv_output
 parts
